Extract commands event and refactor a bit.
