+++ src/events/login.js
@@ -3,7 +3,6 @@
 const util   = require('util');
 const crypto = require('crypto');
 
-
 const EventUtil   = require('./event_util').EventUtil;
 const Data        = require('../data').Data;
 const CommandUtil = require('../command_util').CommandUtil;
@@ -14,9 +13,7 @@ const Commands    = require('../commands').Commands;
 
 const passwordAttempts = {};
 
-console.log(EventUtil);
-
-exports.event = (players, items, rooms, npcs, accounts, l10n) => {
+exports.event =  => {
 
   // Local variables persisted between stages.
   let account = null;
@@ -252,13 +249,12 @@ exports.event = (players, items, rooms, npcs, accounts, l10n) => {
               util.log(player.getName() + ' socket closed during combat!!!');
             }
 
-            //TODO: Consider saving player here as well, and stuff.
             players.removePlayer(player);
           });
 
         //TODO: Have load in player file?
         // Load the player's inventory (There's probably a better place to do this)
-        var inv = [];
+        let inv = [];
         player.getInventory()
           .forEach(item => {
             item = new Item(item);
